requirements.txt
# LangChain Core
langchain
langchain-core

# Environment Variable Management
python-dotenv
streamlit
youtube_transcript_api
pathlib
googletrans

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub